% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{band_plottable_male}
\alias{band_plottable_male}
\title{Non blacklist region for each cytobands binned by 300nt for male.}
\format{
A data frame with columns:
\describe{
  \item{cytoband}{cytoband name (character).}
  \item{coordinates}{ coordinates binned 300nt (integer).}
}
}
\source{
DRAGEN
}
\usage{
band_plottable_male
}
\description{
File used in plot CNV profile to keep only non-blacklist region.
}
\keyword{datasets}
