% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{blacklist_table}
\alias{blacklist_table}
\title{Non mapable regions of the genome}
\format{
A data frame with columns:
\describe{
\item{chrom}{Chromosome name (character).}
\item{start}{Start coordinate (integer).}
\item{end}{End coordinate (integer).}
\item{tag}{ Could be ignored (character).}
}
}
\source{
DRAGEN4.2 CNV output *.cnv.excluded_intervals.bed.gz file
}
\usage{
blacklist_table
}
\description{
A table listing genomic intervals known to be problematic for mapping.
}
\keyword{datasets}
