% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coverage.R
\name{CalbaselineCov}
\alias{CalbaselineCov}
\title{Calculate Baseline Coverage}
\usage{
CalbaselineCov(chrom, cr, count)
}
\arguments{
\item{chrom}{Chromosome}

\item{cr}{Segment mean}

\item{count}{Coverage count}
}
\value{
Baseline coverage (numeric)
}
\description{
Calculate Baseline Coverage
}
