% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modeling.R
\name{Calccf}
\alias{Calccf}
\title{Calculate Cancer Cell Fraction (CCF)}
\usage{
Calccf(CN, mu, rho, C_i)
}
\arguments{
\item{CN}{Numeric. The total copy number in the segment.}

\item{mu}{Numeric. Diploid coverage scale factor (e.g. A value of 1 indicates that the segment mean from GATK does not require adjustment. Additionally, the pseudo-diploid coverage is set to 100.).}

\item{rho}{Numeric. Tumor purity (fraction between 0 and 1).}

\item{C_i}{Numeric. Integer copy number of the mutation.}
}
\value{
Numeric. The estimated cancer cell fraction (CCF), with Inf and -Inf values capped at 1 and 0.01, respectively.
}
\description{
Computes the cancer cell fraction (CCF) based on total copy number,
mutation multiplicity, tumor purity, and integer copy number.
}
\details{
The formula is: \cr
\code{ccf = ((C_i * 2 / (mu * 100)) - ((1 - rho) * 2) - (rho * 2)) / (rho * (CN - 2))} \cr
If CCF is \code{-Inf}, returns 0.01; if \code{Inf}, returns 1; otherwise returns the computed value.
}
\examples{
Calccf(CN = 4, mu = 100, rho = 0.6, C_i = 2)

}
