% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvutils.R
\name{CallwoModel}
\alias{CallwoModel}
\title{Assign Copy Number Call Without Model}
\usage{
CallwoModel(
  chromosome,
  CNF_correct,
  MAF_correct,
  MAF_gmm_G,
  MAF_Probes,
  MAF_gmm_weight,
  opt
)
}
\arguments{
\item{chromosome}{Character. Chromosome name.}

\item{CNF_correct}{Numeric. Corrected CNF.}

\item{MAF_correct}{Numeric. Corrected MAF.}

\item{MAF_gmm_G}{Integer. Number of GMM clusters.}

\item{MAF_Probes}{Integer. Number of probes used for MAF.}

\item{MAF_gmm_weight}{Numeric. GMM weight.}

\item{opt}{List. Must include \code{gender}, \code{callcovcutoff}, \code{callaicutoff}, \code{minsnpcallaicutoff}.}
}
\value{
Character. Copy number call: "GAIN", "LOSS", "REF", "GAINLOH", or "CNLOH".
}
\description{
Assign Copy Number Call Without Model
}
