% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvutils.R
\name{CallWTModel}
\alias{CallWTModel}
\title{Assign Copy Number Call With Model (Major/Minor)}
\usage{
CallWTModel(minor, major, CNF_correct, Chromosome, gender)
}
\arguments{
\item{minor}{Integer. Minor allele copy number.}

\item{major}{Integer. Major allele copy number.}

\item{CNF_correct}{Numeric. Corrected CNF.}

\item{Chromosome}{Character. Chromosome name.}

\item{gender}{Character. "male" or "female".}
}
\value{
Character. Copy number call: "GAIN", "LOSS", "REF", "GAINLOH", or "CNLOH".
}
\description{
Assign Copy Number Call With Model (Major/Minor)
}
