% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coverage.R
\name{CalMedian}
\alias{CalMedian}
\title{Calculate Median Coverage for a Genomic Region}
\usage{
CalMedian(chrom, start, end, cov)
}
\arguments{
\item{chrom}{Chromosome name (character)}

\item{start}{Start position (numeric)}

\item{end}{End position (numeric)}

\item{cov}{Data frame recording coverage with columns CONTIG, START, END, COUNT}
}
\value{
Median coverage (numeric)
}
\description{
Calculate Median Coverage for a Genomic Region
}
