% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIponprocess.R
\name{ChooseNbins}
\alias{ChooseNbins}
\title{Choose Optimal Number of Depth Strata (Bins) for Panel of Normals}
\usage{
ChooseNbins(
  normals_dt,
  target_bins = 8,
  min_rows_per_stratum = 2000,
  min_unique_bins = 100,
  max_bins = 10,
  min_bins = 3
)
}
\arguments{
\item{normals_dt}{Data frame or data.table. Must has depth information of each bin.}
}
\value{
Integer. The chosen number of bins that satisfies the population and diversity constraints.
}
\description{
Determines the optimal number of bins (strata) for grouping samples by SNP median depth, ensuring each bin is sufficiently populated and diverse for downstream modeling.
}
\details{
This function iteratively tests different bin (strata) numbers to find the optimal value based on SNP counts and sequencing depth.
}
