% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVplot.R
\name{CleanHomalt}
\alias{CleanHomalt}
\title{Clean Homozygous AI Bins Based on LOH Segments}
\usage{
CleanHomalt(rounded_ai, call_seg)
}
\arguments{
\item{rounded_ai}{Data frame or tibble. AI data, must include columns \code{contig}, \code{start}, \code{stop}, \code{allele1Count}, \code{allele2Count}.}

\item{call_seg}{Data frame or tibble. Segment calls, must include columns \code{chrom}, \code{loc.start}, \code{loc.end}, and \code{MAF}.}
}
\value{
A data frame of AI bins, with homozygous bins outside LOH regions removed.
}
\description{
Removes homozygous allelic imbalance (AI) bins that do not overlap with LOH (loss of heterozygosity) segments.
}
\examples{
# CleanHomalt(rounded_ai, call_seg)

}
