% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploration_functions.R
\name{cleanKey}
\alias{cleanKey}
\title{Standard cleaning of merge keys}
\usage{
cleanKey(key, also_drop = NULL, to_na = "")
}
\arguments{
\item{key}{Vector of keys to clean; coerced to character.}

\item{also_drop}{Regex pattern specifying additional values to be dropped. Applied piror to \code{to_na}.}

\item{to_na}{Character vector of values to convert to \code{NA}.}
}
\value{
Character vector, a modified version of \code{key}.
}
\description{
Removes all characters from \code{key} \emph{except} for alphanumerics, (internal) space, dash, and underscore. Result is trimmed of leading/trailing white space and converted to uppercase.
Use \code{also_drop} to provide a regex pattern specifying any additional patterns to be dropped. This rule is applied prior to \code{to_na}.
}

