% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{closestMatch}
\alias{closestMatch}
\title{Closest matches by edit distance between all non-matching strings}
\usage{
closestMatch(x, y, collapse = " // ", ignore.case = T)
}
\arguments{
\item{x, y}{Character vectors whose non-identical elements will be compared for closest matches in the other vector.}

\item{collapse}{String used to separate multiple closest-match key values in the case of ties.}

\item{ignore.case}{Logical, if \code{TRUE} (default), case is ignored for computing the distances.}
}
\value{
Dataframe listing the closest matches (and corresponding edit distance) for all non-identical elements in \code{x} and \code{y}.
}
\description{
A typical use case is aligning merge keys when key variables may be fuzzy (e.g., merging pagePath to productName).
}
