% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIsegmentation.R
\name{ClusterAdjacent}
\alias{ClusterAdjacent}
\title{Cluster Adjacent Values Based on a Threshold}
\usage{
ClusterAdjacent(values, weights, threshold = 0.01)
}
\arguments{
\item{values}{Numeric vector of values to cluster (e.g., means from GMM components).}

\item{weights}{Numeric vector of weights corresponding to \code{values}.}

\item{threshold}{Numeric value specifying the maximum allowed difference between adjacent values to be grouped into the same cluster. Default is 0.01.}
}
\value{
A list with two elements:
  \item{means}{Numeric vector of cluster means.}
  \item{weights}{Numeric vector of aggregated cluster weights.}
}
\description{
Groups adjacent values into clusters if their difference is less than a specified threshold,
and computes the mean and total weight for each cluster.
}
\examples{
ClusterAdjacent(c(0.1, 0.11, 0.5), c(0.3, 0.4, 0.3))

}
