% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concat}
\alias{concat}
\alias{\%&\%}
\alias{\%_\%}
\alias{\% \%}
\title{Concatenation Ignoring NA}
\usage{
concat(x, y, sep)

x \%&\% y

x \%_\% y

x \% \% y
}
\arguments{
\item{sep}{a character string to separate the terms.  Not
    \code{\link{NA_character_}}.}
}
\value{
A string equal to the concatenation of \code{...}.
}
\description{
Infix operators for string concatenation. \code{NA} values are ignored rather than pasted as literals (the default \code{paste} behavior).
}
\examples{
"Timestamp: " \%&\% Sys.time()
c(1,2) \%&\% c("a","b")
"foo" \%&\% "bar"
"foo" \%&\% NA
NA \%&\% "bar"
c("foo",NA) \%&\% "buzz"
c("foo","fizz") \% \% c("bar","buzz") \%_\% 1
}
