% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVplot.R
\name{correct_purity}
\alias{correct_purity}
\title{Correct Copy Ratio and Calculate CNF for Purity and Scale Factor}
\usage{
correct_purity(cov, gender, purity, sf)
}
\arguments{
\item{cov}{Data frame or tibble. Coverage data, must include columns \code{CONTIG} and \code{LOG2_COPY_RATIO}.}

\item{gender}{Character. Sample gender ("male" or "female").}

\item{purity}{Numeric. Tumor purity (fraction between 0 and 1).}

\item{sf}{Numeric. Scale factor for diploid coverage.}
}
\value{
The coverage data frame or tibble with additional columns \code{Segcov}, updated \code{LOG2_COPY_RATIO}, and \code{CNF}.
}
\description{
Adjusts the log2 copy ratio and computes the copy number fraction (CNF) for each region, given sample purity and scale factor.
}
\details{
\itemize{
\item Adjusts the log2 copy ratio using the scale factor.
\item Calculates CNF differently for sex chromosomes in males.
}
}
\examples{
# correct_purity(cov, gender = "female", purity = 0.7, sf = 1)

}
