% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIsegmentation.R
\name{CorrectBAF}
\alias{CorrectBAF}
\title{Correct BAF Values Using a Scaling Factor}
\usage{
CorrectBAF(baf, baf_scale)
}
\arguments{
\item{baf}{A data frame or tibble with a numeric column named \code{baf}.}

\item{baf_scale}{Numeric. The scaling factor to be applied to BAF values greater than 0.4.}
}
\value{
A data frame or tibble with the corrected \code{baf} column.
}
\description{
Applies a scaling factor to B-allele frequency (BAF) values greater than 0.4, and caps all BAF values at 0.5.
}
