% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVplot.R
\name{CorrectGender}
\alias{CorrectGender}
\title{Correct Copy Ratio File if GATK and Pipeline Gender Disagree}
\usage{
CorrectGender(cov, seg)
}
\arguments{
\item{cov}{Data frame or tibble. Coverage data, must include columns \code{CONTIG} and \code{LOG2_COPY_RATIO}.}

\item{seg}{Data frame or tibble. Segmentation data, must include columns \code{gatk_gender} and \code{pipeline_gender}.}
}
\value{
The corrected coverage data frame or tibble.
}
\description{
Adjusts the X chromosome log2 copy ratio in the coverage data if the gender assigned by GATK and the pipeline do not match.
}
\details{
- If GATK gender is "male" and pipeline gender is "female", subtracts 1 from X chromosome log2 copy ratio.
- If GATK gender is "female" and pipeline gender is "male", adds 1 to X chromosome log2 copy ratio.
- Otherwise, returns the input coverage unchanged.
}
\examples{
# CorrectGender(cov, seg)

}
