% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIsegmentation.R
\name{CorrectMAF}
\alias{CorrectMAF}
\title{Correct MAF Values Using a Scaling Factor}
\usage{
CorrectMAF(maf, maf_scale)
}
\arguments{
\item{maf}{A data frame or tibble with a numeric column named \code{maf}.}

\item{maf_scale}{Numeric. The scaling factor to be applied to MAF values.}
}
\value{
A data frame or tibble with the corrected \code{maf} column.
}
\description{
Applies a scaling factor to MAF values.
}
