% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CoverageQC.R
\name{CoverageQC}
\alias{CoverageQC}
\title{Summarise BAF QC from CNV Segment Annotation}
\usage{
CoverageQC(annofile)
}
\arguments{
\item{annofile}{Path to the CNV annotation file (e.g., *_CNV_annotation.tsv).}
}
\value{
Invisibly returns the summary data frame.
}
\description{
Summarises the number and size of "PASS" segments per chromosome from a CNV annotation file.
}
