% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytoband_table}
\alias{cytoband_table}
\title{Cytoband reference table}
\format{
A data frame with columns:
\describe{
  \item{chrom}{Chromosome name (character).}
  \item{start}{Start coordinate (integer).}
  \item{end}{End coordinate (integer).}
  \item{band}{Cytoband name (character).}
  \item{gieStain}{Giemsa stain (character).}
}
}
\source{
UCSC Table Browser
}
\usage{
cytoband_table
}
\description{
This dataset contains cytoband annotations for the human genome.
}
\keyword{datasets}
