% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploration_functions.R
\name{deciles}
\alias{deciles}
\alias{icosatiles}
\title{Deciles and icosatiles}
\usage{
deciles(x)

icosatiles(x)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector of specified quantiles.
}
\description{
Deciles and icosatiles
}
\examples{
data(mtcars)
deciles(mtcars$hp)
icosatiles(mtcars$disp)
}

