% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dup.R
\name{dup_}
\alias{dup_}
\alias{dup}
\title{Check dataframe for duplicate keys}
\usage{
dup_(data, ..., .dots, wt = NULL, kable = TRUE)

dup(data, ..., wt = NULL, kable = TRUE)
}
\arguments{
\item{...}{Variable names whose values are to be checked for duplication.}

\item{.dots}{Quoted variable names whose values are to be checked for duplication.}

\item{wt}{Optional weighting variable (as character) used to weight duplication rate diagnostics.}

\item{kable}{Logical, whether to format table for Rmarkdown.}

\item{.data}{The dataframe to check for duplicate cases, where duplication is specific to the variables listed in \code{...}.}
}
\value{
Invisible dataframe of duplicated cases.
}
\description{
Check dataframe for duplicate keys
}
\examples{
d = data.frame(x = rep(1:10, 2), y = rep(1:5, 4), z = 1:20, a = c(1:15, 1:5))
dup(d, y, a)

#SE version
dup_(d, .dots = c("y", "a"))

}
