% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{edge_table}
\alias{edge_table}
\title{Detectable Edge table for cytoband annotation follows the 5MB rule.}
\format{
A data frame with columns:
\describe{
\item{Chrom}{Chromosome name (character).}
\item{p_chromStart}{P arm chromosome Start coordinates (integer).}
\item{p_chromEnd}{P arm chromosome End coordinates (integer).}
\item{p_first_name}{P arm first cytoband name (character).}
\item{p_last_name}{P arm last cytoband name (character).}
\item{q_chromStart}{Q arm chromosome Start coordinates (integer).}
\item{q_chromEnd}{Q arm chromosome End coordinates (integer).}
\item{q_first_name}{Q arm first cytoband name (character).}
\item{q_last_name}{Q arm last cytoband name (character).}
}
}
\source{
Internal analysis
}
\usage{
edge_table
}
\description{
Detectable Edge table for cytoband annotation follows the 5MB rule.
}
\keyword{datasets}
