% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIsegmentation.R
\name{EstimateBAFbyGMM}
\alias{EstimateBAFbyGMM}
\title{Estimate BAF Cluster Mean and Weight Using Gaussian Mixture Modeling}
\usage{
EstimateBAFbyGMM(baf_values)
}
\arguments{
\item{baf_values}{Numeric vector of BAF values.}
}
\value{
Named numeric vector with elements:
\item{gmm_mean}{The mean of the most prominent cluster (capped at 0.5).}
\item{gmm_weight}{The mixture weight of the most prominent cluster.}
\item{gmm_G}{Number of clusters after merging adjacent means.}
}
\description{
Fits a Gaussian mixture model (GMM) to B-allele frequency (BAF) values and returns the most prominent cluster's mean and weight.
Clusters with similar means (within a threshold) are merged. If there are insufficient values or nearly no variance, a simple mean is returned.
}
\details{
Uses \code{\link[mclust]{Mclust}} for Gaussian mixture modeling and \code{\link{ClusterAdjacent}} to merge similar clusters.
}
