% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modeling.R
\name{EstimateSFVariance}
\alias{EstimateSFVariance}
\title{Estimate Standard Deviation of a Numeric Vector}
\usage{
EstimateSFVariance(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
Numeric. The estimated standard deviation of \code{x}.
}
\description{
Calculates the standard deviation (sigma) from the sample variance of a numeric vector.
}
\details{
This function computes the sample variance using \code{var(x)} and returns its square root.
}
\examples{
EstimateSFVariance(c(1, 2, 3, 4, 5))

}
