% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIsegmentation.R
\name{find_peaks}
\alias{find_peaks}
\title{Find the peak values}
\usage{
find_peaks(values, min_valley_drop = 0)
}
\arguments{
\item{values}{any values want to find the peak region}

\item{min_valley_drop}{minimum valley drop for keeping the peaks, default is 0}
}
\value{
peak values and weights
}
\description{
Find the peak values
}
