% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findGrain.R
\name{findGrain_}
\alias{findGrain_}
\alias{findGrain}
\title{Find Data Grain}
\usage{
findGrain_(data, ..., .dots, max_comb = 3, topn = 25)

findGrain(data, ..., max_comb = 3, topn = 25)
}
\arguments{
\item{data}{A dataframe.}

\item{...}{Bare variable names of candidate dimensions to test for defining the grain of \code{data}.}

\item{.dots}{Quoted variable names of candidate dimensions.}

\item{max_comb}{Maximum number of candidate dimensions to combine for testing grain.}

\item{topn}{Number of top combinations to report in results.}
}
\value{
A dataframe with, for each combination of dimensions, the number of duplicate cases (i.e., cases that share the same values of the key variables with one or more other cases). Ideally, if the true grain of the table were found, there would be 0 duplicates.
}
\description{
This function attempts to answer the question: At what grain (combination of dimensions) are records in a dataframe unique?
It searches over all combinations of candidate dimensions (specified via \code{...} or \code{.dots}), up to combinations of size \code{max_comb}, and returns a list of the combinations ranked according to maximum de-duplication of the table.
}
\examples{
data(mtcars)
findGrain(mtcars, gear, cyl, carb)
}
