% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coverage.R
\name{FixsegmentMean}
\alias{FixsegmentMean}
\title{Fix Segment Mean Based on Gender Discrepancy}
\usage{
FixsegmentMean(sm, gatkgender, pipeline_gender)
}
\arguments{
\item{sm}{Segment mean}

\item{gatkgender}{Gender from GATK}

\item{pipeline_gender}{Gender from clinical info}
}
\value{
Adjusted segment mean
}
\description{
Fix Segment Mean Based on Gender Discrepancy
}
