% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_annotation}
\alias{gene_annotation}
\title{RefSeq gene annotations}
\format{
A data frame with columns:
\describe{
\item{chrom}{Chromosome name (character).}
\item{start}{Start coordinate (integer).}
\item{end}{End coordinate (integer).}
\item{genesymbl}{Gene name (character).}
}
}
\source{
RefSeq
}
\usage{
gene_annotation
}
\description{
RefSeq annotations including gene names and start and end.
}
\keyword{datasets}
