% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hasNA}
\alias{hasNA}
\title{Check dataframe for variables with \code{NA}, \code{Inf}, or \code{NaN} values.}
\usage{
hasNA(.data)
}
\arguments{
\item{.data}{The data frame to check.}
}
\value{
\code{summary} of all variables in \code{.data} containing one or more invalid values.
}
\description{
Check dataframe for variables with \code{NA}, \code{Inf}, or \code{NaN} values.
}
\examples{
set.seed(1234)
d = data.frame(w = rnorm(100), x = sample(c(1:10,NA), 100, TRUE), y = c(1:99,Inf), z = c(NaN,2:100), a = sample(LETTERS,100,TRUE))
hasNA(d)
}

