% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVplot.R
\name{keep_whitelist_cov}
\alias{keep_whitelist_cov}
\title{Keep Only Coverage Bins in Whitelisted Regions}
\usage{
keep_whitelist_cov(smooth_cov, whitelist)
}
\arguments{
\item{smooth_cov}{Data frame or tibble. Smoothed coverage data, must include columns \code{CONTIG}, \code{bin_start}, and \code{bin_end}.}

\item{whitelist}{Data frame or tibble. Whitelist regions, must include columns \code{seqnames}, \code{start}, and \code{end}.}
}
\value{
A data frame of smoothed coverage bins that overlap whitelisted regions.
}
\description{
Filters smoothed coverage bins to include only those overlapping whitelisted regions.
}
\examples{
# keep_whitelist_cov(smooth_cov, whitelist)

}
