% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CNVplot.R
\name{KeepWhitelistCov}
\alias{KeepWhitelistCov}
\title{Keep Only Coverage Bins in Whitelisted Regions}
\usage{
KeepWhitelistCov(smooth_cov, whitelist, gender)
}
\arguments{
\item{smooth_cov}{Data frame or tibble. Smoothed coverage data, must include columns \code{CONTIG}, \code{bin_start}, and \code{bin_end}.}

\item{whitelist}{Data frame or tibble. Whitelist regions, must include columns \code{chrom}, \code{start}, and \code{end}.}

\item{gender}{Character. femle or male.}
}
\value{
A data frame of smoothed coverage bins that overlap whitelisted regions.
}
\description{
Filters smoothed coverage bins to include only those overlapping whitelisted regions.
}
\examples{
# KeepWhitelistCov(smooth_cov, whitelist)

}
