% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{mapping}
\alias{mapping}
\title{Mapping of x onto y}
\usage{
mapping(f, data, max_rows = 10, na.rm = FALSE, plot = FALSE,
  kable = FALSE)
}
\arguments{
\item{f}{Formula, \code{x ~ y}}

\item{data}{Dataframe containing \code{x} and \code{y}.}

\item{max_rows}{Maximum number of rows to print (set to \code{NULL} to print all rows).}

\item{na.rm}{Logical, if \code{TRUE} will remove \code{NA} values from \code{x, y}.}

\item{plot}{Logical, if \code{TRUE} will print a bar chart of mapping.}

\item{kable}{Logical, if \code{TRUE} will print table as a \code{kable} for RMarkdown rendering.}
}
\value{
A dataframe summarizing the mapping, in \code{knitr::kable} format if \code{kable = TRUE}. Note that \code{plot} produces a plot as a side-effect.
}
\description{
This function returns the distribution of \eqn{x -> y}, i.e., the number of unique \code{y} values within each unique \code{x} value.
}
\examples{
data(mtcars)
mapping(mpg ~ cyl, mtcars)
}

