% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{matchNames}
\alias{matchNames}
\title{Find names in dataframe matching regex pattern}
\usage{
matchNames(.data, pattern, ignore.case = TRUE)
}
\arguments{
\item{.data}{The data frame to search.}

\item{pattern}{The regex pattern to match against variable names in \code{.data}.}
}
\value{
Character vector of variable names in \code{.data} matching regex \code{pattern}.
}
\description{
Find names in dataframe matching regex pattern
}
\examples{
d = data.frame(x1 = runif(10), y2 = runif(10))
matchNames(d, "x")
d[matchNames(d, "x")]
}

