% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modeling.R
\name{ModelSource}
\alias{ModelSource}
\title{Determine Model Source Based on MAF and Coverage Information}
\usage{
ModelSource(seg_df, modelminprobes)
}
\arguments{
\item{seg_df}{A data frame or tibble containing segment information, including columns: \code{size}, \code{FILTER}, \code{MAF_Probes}, \code{MAF}, and \code{Segment_Mean}.}

\item{modelminprobes}{Numeric. minimum number of probes for a segment to include for model decision}
}
\value{
Character. One of "Coverage", "Diploid", or "Coverage + MAF".
}
\description{
Determines whether the model source for a segment should be "Coverage", "Diploid", or "Coverage + MAF", based on MAF and segment mean values.
}
\details{
\itemize{
\item If all segments with sufficient probes have minimum MAF > 0.4 and the maximum absolute segment mean is >= 0.2, returns "Coverage".
\item If all segments with sufficient probes have minimum MAF > 0.4 and the maximum absolute segment mean is < 0.2, returns "Diploid".
\item Otherwise, returns "Coverage + MAF".
}
}
\examples{
# seg_df must have columns: size, FILTER, MAF_Probes, MAF, Segment_Mean
# ModelSource(seg_df, modelminprobes)

}
