% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIsegmentation.R
\name{ReadAI}
\alias{ReadAI}
\title{Read and Standardize Allelic Imbalance Data}
\usage{
ReadAI(aitype, ballele, minsnpcov, gender)
}
\arguments{
\item{aitype}{Character. Type of AI input file. Supported values are:
\itemize{
\item \code{"gatk"}: Tab-separated file with columns \code{Chromosome}, \code{Pos}, \code{REF_NUCLEOTIDE}, \code{ALT_NUCLEOTIDE}, \code{REF_COUNT}, \code{ALT_COUNT}.
\item \code{"dragen"}: the input must include columns \code{contig}, \code{start}, \code{stop}, \code{refAllele}, \code{allele1}, \code{allele2}, \code{allele1Count}, \code{allele2Count}, \code{allele1AF}, and \code{allele2AF}.
\item \code{"other"}: Tab-separated file with columns \code{Chromosome}, \code{Pos}, \code{REF_NUCLEOTIDE}, \code{ALT_NUCLEOTIDE}, \code{REF_COUNT}, \code{ALT_COUNT}.
}}

\item{ballele}{Character. Path to the input allelic count file.}

\item{minsnpcov}{Numeric, minimum coverage of SNPs to included (default: 20).}

\item{gender}{Character. Sample gender, either \code{"female"} or \code{"male"}. This value is passed to \code{ReadAI()}.}
}
\value{
A data frame with columns \code{Chromosome}, \code{Pos}, \code{ref_count}, \code{alt_count} and \code{maf}, containing only autosomal positions (\code{Chromosome} not X or Y), with minor allele frequency (\code{maf}) between 0 and 0.5.
}
\description{
Reads allelic imbalance (AI) data from different input formats and standardizes the output.
}
\examples{
\dontrun{
# Read AI data from a count file
maf_df <- ReadAI("gatk", "sample.allele_counts", 20, "female")

# Read AI data from a DRAGEN file
maf_df <- ReadAI("dragen", "sample_dragen.tumor.ballele.counts.gz",20, "female")

# Read AI data from a BAF file
maf_df <- ReadAI("other", "sample.tsv",20, "female")
}

}
