% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modeling.R
\name{RefineCalls}
\alias{RefineCalls}
\title{Refine Segment Calls According to the Final Model}
\usage{
RefineCalls(df, max_L_mu, max_L_rho, gender)
}
\arguments{
\item{df}{Data frame or tibble. Segment call data, must include columns used in refinement.}

\item{max_L_mu}{Numeric. The selected value of mutation multiplicity (\code{mu}).}

\item{max_L_rho}{Numeric. The selected value of tumor purity (\code{rho}).}

\item{gender}{Character. Gender female or male.}
}
\value{
A data frame or tibble with refined segment calls, ordered by chromosome and start position.
}
\description{
Refines values and calls for each segment based on the selected (mu, rho) model, adjusting segment mean, purity, and copy number calls.
}
\examples{
# RefineCalls(df, max_L_mu = 1, max_L_rho = 0.7, gender = "female")

}
