% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modeling.R
\name{RefineCallsSecond}
\alias{RefineCallsSecond}
\title{Second Refinement of Segment Calls for Cov/AI Profile Mismatches}
\usage{
RefineCallsSecond(df, results, final_mu, final_rho, opt)
}
\arguments{
\item{df}{Data frame or tibble. Refined segment calls (see \code{\link{RefineCalls}}), must include columns as in \code{col_name}.}

\item{results}{Data frame or tibble. All likelihood results for possible copy number states, must include columns \code{rho}, \code{mu}, \code{index}, \code{total_cn}, etc.}

\item{final_mu}{Numeric. The selected value of mutation multiplicity (\code{mu}).}

\item{final_rho}{Numeric. The selected value of tumor purity (\code{rho}).}

\item{opt}{List of options, must include \code{gender} and \code{callcov}.}
}
\value{
A data frame or tibble with further refined segment calls, including new columns \code{ccf_COV}, \code{ccf_final}, and \code{CN_mix}.
}
\description{
Further refines segment calls for segments where the coverage profile and allelic imbalance (AI) profile do not match,
by considering alternative copy number states from the likelihood results.
}
\examples{
# RefineCallsSecond(df, results, final_mu = 1, final_rho = 0.7, opt)

}
