% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnvutils.R
\name{roundcn}
\alias{roundcn}
\title{Round Copy Number to Integer Based on Call and Gender}
\usage{
roundcn(gender, Chrom, Call, CNF)
}
\arguments{
\item{gender}{Character. "male" or "female".}

\item{Chrom}{Character. Chromosome name.}

\item{Call}{Character. CNV call.}

\item{CNF}{Numeric. Corrected CNF.}
}
\value{
Integer. Rounded copy number.
}
\description{
Round Copy Number to Integer Based on Call and Gender
}
