% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAIsegmentation.R
\name{RunAIsegmentation}
\alias{RunAIsegmentation}
\title{Run AI Segmentation Workflow}
\usage{
RunAIsegmentation(
  seg,
  cov,
  cr,
  ai,
  gender,
  out_dir,
  prefix,
  ai_pon,
  sampletype,
  covminvalleydrop = 0.1,
  covminproportion = 0.2,
  covminpeakdiff = 0.5,
  bafminproportion = 0.25,
  mergeai = 0.05,
  mergecov = 0.05,
  snpmin = 4,
  minsnpcov = 20,
  mergecovminsize = 5e+05,
  maxgap = 2e+06,
  snpnum = 20,
  maxbinsize = 1e+06,
  minbinsize = 5e+05,
  mafqcweightcutoff = 0.4,
  cov_mad_cutoff1 = 0.6,
  cov_mad_cutoff2 = 0.3,
  aitype = "gatk",
  segmethod = "cbs",
  cbssmooth = "no"
)
}
\arguments{
\item{seg}{Path to the GATK segment file.}

\item{cov}{Path to the GATK denoised coverage count file.}

\item{cr}{Path to the GATK denoised coverage ratio file.}

\item{ai}{Path to the BAF file or allelic count file.}

\item{gender}{Character, "male" or "female"( if not available, set gender = "unknown" and run Runcheckgender function ahead to assign gender automatically) .}

\item{out_dir}{Output directory path.}

\item{prefix}{Output file prefix.}

\item{ai_pon}{Path to PON Rdata. AI panel of normal generated by \code{PONAIprocess}.}

\item{sampletype}{Character, "ff" or "ffpe".}

\item{covminvalleydrop}{minimum valley drop in the finding peaks of coverage track quality control, default is 0.1}

\item{covminproportion}{minimum proportion of peak in the finding peaks of coverage track quality control, default is 0.2}

\item{covminpeakdiff}{minimum peak difference in the finding peaks of coverage track quality control, default is 0.5}

\item{bafminproportion}{Minimum proportion of SNPs required for a BAF peak
to be retained for MAF estimation; smaller peaks are excluded. default is 0.25}

\item{mergeai}{Numeric, MAF difference threshold for merging segments (default: 0.05).}

\item{mergecov}{Numeric, CNV difference threshold for merging segments (default: 0.05).}

\item{snpmin}{Numeric, minimum SNPs for MAF segmentation (default: 4).}

\item{minsnpcov}{Numeric, minimum coverage of SNPs to included (default: 20).}

\item{mergecovminsize}{Numeric, minimum size for GATK segment merge (default: 500000).}

\item{maxgap}{Maximum gap size inside a bin. If exceed then start another bin.( default: 2000000)}

\item{snpnum}{SNP number in each bin.( default: 20 )}

\item{maxbinsize}{Maximum bin size.( default: 1000000 )}

\item{minbinsize}{Minimum bin size.( default: 500000 )}

\item{mafqcweightcutoff}{Numeric. Minimum SNP proportion in the main MAF peak
required to pass MAF QC. Values below this cutoff fail QC. Default is 0.4.}

\item{cov_mad_cutoff1}{Numeric. Over this value will directly Fail qc. default is 0.6}

\item{cov_mad_cutoff2}{Numeric. Over this value, but lower than cov_mad_cutoff1, if multiple peaks will fail the qc. default is 0.3}

\item{aitype}{Character. Type of allelic imbalance data: \code{"gatk"} or \code{"other"} or \code{"dragen"}
If \code{"GATK"}, the input must include columns \code{CONTIG}, \code{POSITION}, \code{ALT_COUNT}, \code{REF_COUNT}, \code{REF_NUCLEOTIDE}, and \code{ALT_NUCLEOTIDE}.
If \code{"dragen"}, the input must include columns \code{contig}, \code{start}, \code{stop}, \code{refAllele}, \code{allele1}, \code{allele2}, \code{allele1Count}, \code{allele2Count}, \code{allele1AF}, and \code{allele2AF}.
If \code{"other"}, the input must include columns \code{CONTIG}, \code{POSITION}, \code{ALT_COUNT}, \code{REF_COUNT}, \code{REF_NUCLEOTIDE}, and \code{ALT_NUCLEOTIDE}.}

\item{segmethod}{Character. Segmentation method to use: if \code{"merge"}, perform stepwise merging; if \code{"cbs"}, perform CBS (circular binary segmentation).}

\item{cbssmooth}{Character. If using the \code{"cbs"} segmentation method, set to \code{"yes"} to apply smoothing before segmentation, or \code{"no"} to skip smoothing.}
}
\value{
Invisibly returns the output file path.
}
\description{
This function executes the full AI segmentation workflow, including reading input files, merging segments, BAF processing, and writing results.
}
