% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runModelLikelihood.R
\name{RunModelLikelihood}
\alias{RunModelLikelihood}
\title{Run Model Likelihood and Copy Number Calling Workflow}
\usage{
RunModelLikelihood(
  seg,
  out_dir,
  prefix,
  gender,
  lambda = 1,
  gamma = 1,
  epsilon = 0.01,
  modelminprobes = 20,
  modelminAIsize = 5e+06,
  minsf = 0.4,
  callcov = 0.3,
  thread = 4,
  XploR_path = NULL,
  ...
)
}
\arguments{
\item{seg}{Path to the combined segment file.}

\item{out_dir}{Output directory.}

\item{prefix}{Output file prefix.}

\item{gender}{Sample gender ("male" or "female").}

\item{modelminprobes, }{modelminAIsize, callcov, thread, callcovcutoff, callaicutoff, minsnpcallaicutoff Numeric parameters for the workflow (see script RunCallikelihood for details).}

\item{XploR_path, }{Path of XploR package, if don't want to install.}
}
\value{
Invisibly returns the output file paths.
}
\description{
This function runs the complete workflow for likelihood calculation and model selection.
}
