% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPlotCNV.R
\name{runPlotCNV}
\alias{runPlotCNV}
\title{Plot CNV Profile (Copy Number, BAF, Clonality, Quality)}
\usage{
runPlotCNV(
  seg,
  coverage,
  ballele,
  ai_binsize = 1e+05,
  cov_binsize = 1e+05,
  whitelist,
  gender = "female",
  out_dir,
  prefix
)
}
\arguments{
\item{seg}{Path to the GATK-processed segment file.}

\item{coverage}{Path to the coverage file.}

\item{ballele}{Path to the DRAGEN B-allele file.}

\item{ai_binsize}{Numeric. Bin size for AI plot (default 100000).}

\item{cov_binsize}{Numeric. Bin size for coverage plot (default 100000).}

\item{whitelist}{Path to whitelist file.}

\item{gender}{Character. "male" or "female".}

\item{out_dir}{Output directory.}

\item{prefix}{Sample ID or output prefix.}
}
\value{
Invisibly returns the ggplot object.
}
\description{
Plots the CNV profile using coverage, B-allele, segment, and whitelist data.
}
