% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIsegmentation.R
\name{SearchBreakpoint}
\alias{SearchBreakpoint}
\title{Refine Breakpoints Within a Segment Using MAF Data}
\usage{
SearchBreakpoint(
  seg_row,
  maf,
  pon_ref,
  gender,
  out_dir,
  prefix,
  mergeai = 0.05,
  snpmin = 3,
  maxgap = 1e+06,
  snpnum = 20,
  maxbinsize = 1e+06,
  minbinsize = 5e+05,
  minsnpcov = 20,
  bafminseparationsd = 1.1,
  bafminproportion = 0.25,
  segmethod = "cbs",
  cbssmooth = "yes",
  covmad_cutoff,
  sampletype
)
}
\arguments{
\item{seg_row}{Data frame row (list or tibble row) representing a single segment. Must have columns: Sample, Chromosome, Start, End, Num_Probes, Segment_Mean, Segment_Mean_raw, Count, Baseline_cov, gatk_gender, pipeline_gender, size.}

\item{maf}{Data frame or tibble containing MAF data. Must include columns: Chromosome, Pos, maf.}

\item{pon_ref}{Data frame. Panel of normal reference for bias correction (required for bias correction step).}

\item{gender}{Character. If \code{"female"}, the X chromosome will also be proceed.}

\item{mergeai}{Numeric. Threshold for the difference in MAF (gmm_mean) between adjacent segments to allow merging under \code{"merge"} mode segmentation.}

\item{snpmin}{Numeric. Minimum SNP count required for a segment to be considered as a separate segment under \code{"merge"} mode segmentation.}

\item{maxgap}{Numeric. Maximum allowed gap between SNPs within a bin.}

\item{snpnum}{Integer. Target number of SNPs per bin.}

\item{maxbinsize}{Numeric. Maximum allowed bin size (bp).}

\item{minbinsize}{Numeric. Minimum allowed bin size (bp). The minimum segment size under \code{"merge"} mode is 2*minbinsize.}

\item{minsnpcov}{Integer. Minimum coverage of SNP sites to be included.}

\item{bafminseparationsd}{Minimum peak separation in units of fitted SD. Default 1.1.}

\item{bafminproportion}{Minimum proportion of SNPs required for a BAF peak
to be retained for MAF estimation; smaller peaks are excluded. default is 0.25}

\item{segmethod}{Character. Segmentation method to use: if \code{"merge"}, perform stepwise merging; if \code{"cbs"}, perform CBS (circular binary segmentation).}

\item{cbssmooth}{Character. If using the \code{"cbs"} segmentation method, set to \code{"yes"} to apply smoothing before segmentation, or \code{"no"} to skip smoothing.}
}
\value{
A data frame with the refined segment(s), including updated breakpoints, MAF metrics, and a \code{BreakpointSource} column indicating whether breakpoints were post-processed or from GATK.
}
\description{
Refines breakpoints within a segment using minor allele frequency (MAF) data. If enough informative MAF sites are present, the segment is binned and can be split into finer regions using either stepwise merging or CBS (circular binary segmentation). Optionally, PON-based bias correction is applied to the resulting segments.
}
\details{
The function first bins the MAF data within the segment. If \code{segmethod = "merge"}, segments are merged stepwise based on the MAF difference and SNP count. If \code{segmethod = "cbs"}, CBS segmentation is performed on the binned MAF values, with optional smoothing. After segmentation, bias correction using the panel of normal can be applied. The function returns refined segments with updated metrics and a \code{BreakpointSource} label.
}
