% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/joins.R
\name{vjoin}
\alias{vjoin}
\title{Verbose Join}
\usage{
vjoin(x, y, by, join = "left", flag_source = FALSE)
}
\arguments{
\item{x, y}{Dataframes to join.}

\item{by}{Character vector giving the name(s) of columns to be used for merging.}

\item{join}{String specifying type of join: valid join types are "inner", "left", "right", "full", "semi" and "anti".}

\item{flag_source}{Logical, whether to add two variables (\code{.in_x} and \code{.in_y}) flagging source(s) for each case in result.}
}
\value{
Merged data frame.
}
\description{
Join with diagnostic output. Joins are implemented via \code{dplyr} join functions.
}
\examples{
x <- data.frame(a = LETTERS[1:10], b = 1:10, stringsAsFactors = F)
y <- data.frame(a = LETTERS[6:15], c = 6:15, stringsAsFactors = F)
vjoin(x, y, by = "a", join = "full")
}

