% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BafQC.R
\name{BafQC}
\alias{BafQC}
\title{Summarise BAF QC from CNV Segment Annotation}
\usage{
BafQC(annofile, out_dir, prefix)
}
\arguments{
\item{annofile}{Path to the CNV annotation file (e.g., *_CNV_annotation.tsv).}

\item{out_dir}{Path to QC output file.}

\item{prefix}{Prefix of QC output file.}
}
\value{
Invisibly returns the summary data frame.
}
\description{
Summarises the number and size of "PASS" segments per chromosome from a CNV annotation file.
}
