% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AIsegmentation.R
\name{BinMaf}
\alias{BinMaf}
\title{Bin MAF Data and Estimate MAF Metrics per Bin}
\usage{
BinMaf(
  data,
  datatype,
  maxgap = 2e+06,
  snpnum = 20,
  maxbinsize = 1e+06,
  minbinsize = 5e+05,
  minsnpcov = 20
)
}
\arguments{
\item{data}{A data frame or tibble containing at least \code{Chromosome}, \code{Pos}, \code{Start}, \code{End} and \code{maf} columns.}

\item{datatype}{If individual tumor data then choose "tumor", if aggregated panel of normal samples choose "pon".}

\item{maxgap}{Maximum gap size inside a bin. If exceed then start another bin.( default: 2000000)}

\item{snpnum}{SNP number in each bin.( default: 20 )}

\item{maxbinsize}{Maximum bin size.( default: 1000000 )}

\item{minbinsize}{Minimum bin size.( default: 500000 )}

\item{minsnpcov}{Minimum coverage of SNP sites to be included. ( default: 20 )}
}
\value{
A tibble with one row per bin and columns:
\item{Chromosome}{Chromosome identifier.}
\item{bin}{Bin start coordinate.}
\item{Start}{Minimum \code{End} value in the bin.}
\item{End}{Maximum \code{End} value in the bin.}
\item{nonzero_count}{Number of nonzero maf values in the bin.}
\item{each_maf}{Semicolon-separated string of nonzero maf values in the bin.}
\item{gmm_mean, gmm_weight, gmm_G}{Unnested maf GMM metrics.}
}
\description{
Bins input MAF data into fixed-size windows or fixed-snp-count window and calculates summary metrics (e.g., estimated MAF, nonzero counts) for each bin.
}
