% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modeling.R
\name{CalSegmentLikelihood}
\alias{CalSegmentLikelihood}
\title{Calculate Marginal Likelihood for a Single Segment}
\usage{
CalSegmentLikelihood(C_i, B_i, mu, rho, sigma_C, k, lambda, gamma, epsilon)
}
\arguments{
\item{C_i}{Numeric. Integer copy number for the segment.}

\item{B_i}{Numeric. Observed B-allele frequency (MAF) for the segment.}

\item{mu}{Numeric. Diploid coverage scale factor (e.g. A value of 1 indicates that the segment mean from GATK does not require adjustment. Additionally, the pseudo-diploid coverage is set to 100.).}

\item{rho}{Numeric. Tumor purity (fraction between 0 and 1).}

\item{sigma_C}{Numeric. Not directly used in this function, but included for compatibility.}

\item{k}{Numeric. Beta distribution concentration parameter.}

\item{lambda}{Numeric. Exponential decay parameter for the prior.}

\item{gamma}{Numeric. Weight for the prior in the likelihood calculation.}

\item{epsilon}{Numeric. Small value to avoid log(0) and zero parameters in beta.}
}
\value{
A data frame with columns: major, minor, CN, ccf, Bio_diff, prior, expected_MAF, MAF_ll, weighted_prior, exp_MAF_ll, exp_prior, MAF_likelihood, Segcov, MAF, mu, rho.
}
\description{
Computes the marginal likelihood for a segment given copy number, MAF, purity, and model parameters.
}
\details{
The returned data frame contains the following columns:
\itemize{
\item major: Major allele count
\item minor: Minor allele count
\item CN: Total copy number
\item ccf: Cancer cell fraction
\item Bio_diff: Biological difference
\item prior: Prior probability
\item expected_MAF: Expected MAF
\item MAF_ll: MAF log-likelihood
\item weighted_prior: Weighted prior
\item exp_MAF_ll: Expected MAF log-likelihood
\item exp_prior: Expected prior
\item MAF_likelihood: MAF likelihood
\item Segcov: Segment coverage
\item MAF: Observed MAF
\item mu: Mutation multiplicity
\item rho: Tumor purity
}
}
