% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coverage.R
\name{CheckGender}
\alias{CheckGender}
\title{Check and Adjust Gender in GATK Segmentation Data}
\usage{
CheckGender(cov, seg, gender)
}
\arguments{
\item{cov}{Coverage data frame}

\item{seg}{Segmentation data frame}

\item{gender}{Gender from clinical information (character, "male", "female" or "unknown")}
}
\value{
Modified segmentation seg data frame with gender information
}
\description{
Check and Adjust Gender in GATK Segmentation Data
}
