% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rerun.R
\name{Checkmode}
\alias{Checkmode}
\title{Check and Set Mode-Specific Options}
\usage{
Checkmode(opt, call)
}
\arguments{
\item{opt}{List of options. Must include \code{mode} (either "model" or "region"), and possibly \code{purity}, \code{dicovsf}, \code{chromosome}.}

\item{call}{Data frame or tibble. Segment call data, required if estimating coverage scale factor for region mode.}
}
\value{
The (potentially modified) \code{opt} list.
}
\description{
Checks required options for 'model' or 'region' mode and sets defaults if needed.
}
\details{
\itemize{
\item If \code{mode == "model"}: requires at least one of \code{purity} or \code{dicovsf}.
\item If \code{mode == "region"}: requires \code{chromosome}; if \code{dicovsf} is missing, it is estimated from \code{call}.
}
}
\examples{
# opt <- list(mode = "model", purity = NULL, dicovsf = NULL)
# Checkmode(opt, call)

}
