% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpf.R
\name{cpf_}
\alias{cpf}
\alias{cpf_}
\alias{has}
\alias{has_}
\title{\strong{C}umulative \strong{P}ercentile \strong{F}requency Table}
\usage{
cpf_(data, ..., .dots, wt = NULL, sort = TRUE, margin = TRUE,
  chi_square = FALSE, kable = FALSE)

cpf(data, ..., wt = NULL, sort = TRUE, margin = TRUE,
  chi_square = FALSE, kable = FALSE)

has(data, ..., wt = NULL, sort = TRUE, margin = TRUE,
  chi_square = FALSE, kable = FALSE)

has_(data, ..., .dots, wt = NULL, sort = TRUE, margin = TRUE,
  chi_square = FALSE, kable = FALSE)
}
\arguments{
\item{data}{A dataframe.}

\item{...}{Variable names defining table groupings.}

\item{.dots}{Vector of variable names defining table groupings.}

\item{wt}{(Optional, character) variable name to use for weighting frequencies.}

\item{sort}{Logical, whether to sort table in descending frequency.}

\item{margin}{Logical, whether to add total row to end of table.}

\item{chi_square}{Logical, whether to print chi-square test (has no effect on one-dimensional tables).}

\item{kable}{Logical, whether to format table for Rmarkdown.}
}
\value{
A dataframe or \code{kable}.
}
\description{
\strong{C}umulative \strong{P}ercentile \strong{F}requency Table
}
\examples{
data(mtcars)
cpf(mtcars, cyl, gear)
}

