% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpf.R
\name{cpf_}
\alias{cpf}
\alias{cpf_}
\title{\strong{C}umulative \strong{P}ercentile \strong{F}requency Table}
\usage{
cpf_(data, ..., .dots, wt = NULL, sort = TRUE, kable = FALSE)

cpf(data, ..., wt = NULL, sort = TRUE, kable = FALSE)
}
\arguments{
\item{data}{A dataframe.}

\item{...}{Bare variable names defining table groupings.}

\item{.dots}{Quoted variable names defining table groupings.}

\item{wt}{(Optional) quoted variable name to use for weighting frequencies.}

\item{sort}{Logical, whether to sort table in descending frequency.}

\item{kable}{Logical, whether to format table for Rmarkdown.}
}
\value{
A dataframe or \code{kable}.
}
\description{
\strong{C}umulative \strong{P}ercentile \strong{F}requency Table
}
\examples{
data(mtcars)
cpf(mtcars, cyl, gear)
}

