% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploration_functions.R
\name{cpf}
\alias{cpf}
\alias{cpf_}
\title{Cumulative Percentile/Frequency Table}
\usage{
cpf(data, ..., wt = NULL, sort = TRUE, kable = FALSE)

cpf_(data, ..., .dots, wt = NULL, sort = TRUE, kable = FALSE)
}
\arguments{
\item{data}{A dataframe.}

\item{...}{Bare variable names defining table groupings.}

\item{wt}{(Optional) quoted variable name to use for weighting frequencies.}

\item{sort}{Logical, whether to sort table in descending frequency.}

\item{kable}{Logical, whether to format table for Rmarkdown.}

\item{.dots}{Quoted variable names defining table groupings.}
}
\value{
A dataframe or \code{kable}.
}
\description{
Cumulative Percentile/Frequency Table
}
\examples{
data(mtcars)
cpf(mtcars, cyl, gear)
}

