% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpf.R
\name{cpf_}
\alias{cpf}
\alias{cpf_}
\alias{has}
\alias{has_}
\title{\strong{C}umulative \strong{P}ercentile \strong{F}requency Table}
\usage{
cpf_(data, ..., .dots, wt = NULL, sort = TRUE, margin = TRUE,
  chi_square = FALSE, kable = FALSE)

cpf(data, ..., wt = NULL, sort = TRUE, margin = TRUE,
  chi_square = FALSE, kable = FALSE)

has(data, ..., wt = NULL, sort = TRUE, margin = TRUE,
  chi_square = FALSE, kable = FALSE)

has_(data, ..., .dots, wt = NULL, sort = TRUE, margin = TRUE,
  chi_square = FALSE, kable = FALSE)
}
\arguments{
\item{data}{A dataframe.}

\item{...}{Variable names defining table groupings.}

\item{.dots}{Vector of variable names defining table groupings.}

\item{wt}{(Optional, character) variable name to use for weighting frequencies.}

\item{sort}{Logical, whether to sort table in descending frequency. Setting to \code{FALSE} will sort table alphabetically by grouping variables.}

\item{margin}{Logical, whether to add total row to end of table.}

\item{chi_square}{Logical, whether to print chi-square test (has no effect on one-dimensional tables).}

\item{kable}{Logical, whether to format table for Rmarkdown.}
}
\value{
A dataframe or \code{kable}.
}
\description{
\code{cpf} produces a frequency table by the specified variables, which should be categorical and of relatively low cardinality.
}
\details{
\code{has} is a wrapper around \code{cpf} for the common pattern of counting the number of \code{NA} and non-\code{NA} values in each specified variable. It is equivalent to \code{cpf(data, !is.na(x), !is.na(y), ...)}.
}
\examples{
data(mtcars)
cpf(mtcars, cyl, gear)
}

