% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{deciles}
\alias{deciles}
\alias{icosatiles}
\title{Deciles and icosatiles}
\usage{
deciles(x, na.rm = TRUE)

icosatiles(x, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric vector}
}
\value{
Numeric vector of specified quantiles.
}
\description{
Deciles and icosatiles
}
\examples{
data(mtcars)
deciles(mtcars$hp)
icosatiles(mtcars$disp)
}

