% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exploration_functions.R
\name{dup}
\alias{dup}
\alias{dup_}
\title{Check data frame for duplicates by index(es).}
\usage{
dup(data, ..., verbose = FALSE)

dup_(data, ..., .dots, verbose = FALSE)
}
\arguments{
\item{...}{Bare variable names whose values are to be checked for duplication.}

\item{verbose}{If \code{TRUE}, will print all duplicated cases in \code{data}.}

\item{.dots}{Quoted variable names whose values are to be checked for duplication.}

\item{.data}{The data frame to check for duplicate cases, where duplication is specific to the variables listed in \code{...}.}
}
\value{
Invisible dataframe of duplicated cases.
}
\description{
Check data frame for duplicates by index(es).
}
\examples{
d = data.frame(x = rep(1:10, 2), y = rep(1:5, 4), z = 1:20, a = c(1:15, 1:5))
dup(d, y, a, verbose = T)

#SE version
dup_(d, .dots = c("y", "a"))

}

