% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coverage.R
\name{MergeSegRow}
\alias{MergeSegRow}
\title{Merge Adjacent Copy Number Segments Based on Similarity Criteria}
\usage{
MergeSegRow(df, mergecov)
}
\arguments{
\item{df}{A data frame or tibble of copy number segments, with columns such as Sample, Chromosome, Start, End, Num_Probes, Call, Segment_Mean, etc.}

\item{mergecov}{Numeric threshold for merging of segment mean value.}
}
\value{
A data frame or tibble with merged segments and updated segment information.
}
\description{
Iteratively merges adjacent rows in a data frame of copy number segments if they meet criteria defined by \code{MergeSegCheck}.
For merged segments, the segment with the greater number of probes provides the call, segment mean, count, and baseline coverage.
}
\details{
This function uses \code{\link{MergeSegCheck}} to determine if two adjacent segments should be merged.
}
