% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAIsegmentation.R
\name{RunAIsegmentation}
\alias{RunAIsegmentation}
\title{Run AI Segmentation Workflow}
\usage{
RunAIsegmentation(
  seg,
  cov,
  ai,
  gender,
  out_dir,
  prefix,
  aibinsize = 5e+05,
  mergeai = 0.2,
  mergecov = 0.2,
  minaisize = 1e+06,
  snpmin = 7,
  minsnpcallaicutoff = 10,
  mergecovminsize = 5e+05,
  aitype = "count"
)
}
\arguments{
\item{seg}{Path to the GATK segment file.}

\item{cov}{Path to the GATK denoised coverage count file.}

\item{ai}{Path to the BAF file or allelic count file.}

\item{gender}{Character, "male" or "female".}

\item{out_dir}{Output directory path.}

\item{prefix}{Output file prefix.}

\item{aibinsize}{Numeric, AI bin size (default: 500000).}

\item{mergeai}{Numeric, MAF difference threshold for merging segments (default: 0.15).}

\item{mergecov}{Numeric, CNV difference threshold for merging segments (default: 0.2).}

\item{minaisize}{Numeric, smallest AI segment size (default: 1000000).}

\item{snpmin}{Numeric, minimum SNPs for MAF segmentation (default: 7).}

\item{minsnpcallaicutoff}{Numeric, minimum SNPs for reliable CNLOH/GAINLOH (default: 10).}

\item{mergecovminsize}{Numeric, minimum size for GATK segment merge (default: 500000).}

\item{aitype}{Character. Type of allelic imbalance data: \code{"gatk"} or \code{"other"} or \code{"dragen"}
If \code{"GATK"}, the input must include columns \code{CONTIG}, \code{POSITION}, \code{ALT_COUNT}, \code{REF_COUNT}, \code{REF_NUCLEOTIDE}, and \code{ALT_NUCLEOTIDE}.
If \code{"dragen"}, the input must include columns \code{contig}, \code{start}, \code{stop}, \code{refAllele}, \code{allele1}, \code{allele2}, \code{allele1Count}, \code{allele2Count}, \code{allele1AF}, and \code{allele2AF}.
If \code{"other"}, the input must include columns \code{CONTIG}, \code{POSITION}, \code{ALT_COUNT}, \code{REF_COUNT}, \code{REF_NUCLEOTIDE}, and \code{ALT_NUCLEOTIDE}.}
}
\value{
Invisibly returns the output file path.
}
\description{
This function executes the full AI segmentation workflow, including reading input files, merging segments, BAF processing, and writing results.
}
