% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runPlotCNV.R
\name{RunPlotCNV}
\alias{RunPlotCNV}
\title{Plot CNV Profile (Copy Number, BAF, Clonality, Quality)}
\usage{
RunPlotCNV(
  seg,
  cr,
  ballele,
  ai_binsize = 1e+05,
  cov_binsize = 1e+05,
  whitelist,
  gender = NULL,
  out_dir,
  prefix,
  aitype
)
}
\arguments{
\item{seg}{Path to the GATK-processed segment file.}

\item{cr}{Path to the GATK denoised CR file with extension denoisedCR.tsv.}

\item{ballele}{Path to the DRAGEN B-allele file. Or other souce B-allele file with column chrom ref alt baf}

\item{ai_binsize}{Numeric. Bin size for AI plot (default 100000).}

\item{cov_binsize}{Numeric. Bin size for coverage plot (default 100000).}

\item{whitelist}{Path to whitelist file.}

\item{gender}{Character. "male" or "female".}

\item{out_dir}{Output directory.}

\item{prefix}{Sample ID or output prefix.}

\item{aitype}{Character. Type of allelic imbalance data: \code{"gatk"} or \code{"other"} or \code{"dragen"}
If \code{"GATK"}, the input must include columns \code{CONTIG}, \code{POSITION}, \code{ALT_COUNT}, \code{REF_COUNT}, \code{REF_NUCLEOTIDE}, and \code{ALT_NUCLEOTIDE}.
If \code{"dragen"}, the input must include columns \code{contig}, \code{start}, \code{stop}, \code{refAllele}, \code{allele1}, \code{allele2}, \code{allele1Count}, \code{allele2Count}, \code{allele1AF}, and \code{allele2AF}.
If \code{"other"}, the input must include columns \code{CONTIG}, \code{POSITION}, \code{ALT_COUNT}, \code{REF_COUNT}, \code{REF_NUCLEOTIDE}, and \code{ALT_NUCLEOTIDE}.}
}
\value{
Invisibly returns the ggplot object.
}
\description{
Plots the CNV profile using coverage, B-allele, segment, and whitelist data.
}
