% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ss.R
\name{ss_}
\alias{ss_}
\alias{ss}
\title{\strong{S}ummary \strong{S}tatistics Table}
\usage{
ss_(data, ..., .dots, funs = summary_funs(), kable = FALSE, digits = 2,
  plot = FALSE)

ss(data, ..., funs = summary_funs(), kable = FALSE, digits = 2,
  plot = FALSE)
}
\arguments{
\item{data}{A dataframe.}

\item{...}{Bare names of numeric variables to summarize (NSE); if empty will summarize all numeric columns in \code{data}.}

\item{.dots}{Quoted names of numeric variables to summarize (SE).}

\item{funs}{List of summary functions defined as formulas on \code{x}.}

\item{kable}{Logical, whether to return a \code{kable} for rendering in Rmarkdown.}

\item{digits}{Number of digits to print after decimal.}

\item{plot}{Logical, whether to plot (as a side-effect) a facet-wrapped set of histograms for all summarized variables.}
}
\value{
A data.frame or kable object containing the table of summary statistics.
}
\description{
Produce a compact table of summary statistics for all or specified numeric variables in a dataframe.
}
